fix(discovery): resolve installation relationship + extract ingest core from route
Two QA findings: the ingest route wrote the numeric GitHub installation id
into the installation relationship (Mongo ObjectId cast 500 on every real
scan — FakePayload couldn't catch it; now resolved via github-installations
lookup, with regression tests), and Next's generated route types reject
non-handler exports, so ingestScan/parseBody move to lib/discovery/ingest.ts.

Verified live against the dev stack: tier-1 .orbit.yaml auto-import creates
the App and projected catalog entity; double-POST creates no duplicate rows.
